requirements.txt
asgiref==3.3.1
assertpy==1.1
astroid==2.4.2
awsebcli==3.19.4
botocore==1.20.69
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.3
django-ajax==0.3.0
django-allauth==0.44.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy==1.8.1
django-debug-toolbar==3.2
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
future==0.16.0
gunicorn==20.0.4
idna==2.10
importlib==1.0.4
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
oauthlib==3.1.1
pathspec==0.5.9
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.1
PyJWT==2.1.0
pylint==2.6.0
pypiwin32==223
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
pywin32==300
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
urllib3==1.25.11
wcwidth==0.1.9
whitenoise==5.2.0
wrapt==1.12.1
